:recycle:refactor: refactor feed post card 컴포넌트, 스타일 분리 Related to: #78
