+++ src/components/FeedPage/PostCard/PostCard.styles.ts
@@ -1,0 +1,157 @@
+export const styles = {
+  card: {
+    display: 'flex',
+    flexDirection: 'column',
+    boxSizing: 'border-box',
+    width: '100%',
+    aspectRatio: 'auto',
+    bgcolor: '#fafafa',
+    borderRadius: '8px',
+    transition: 'all 0.3s ease-in-out',
+    '&:hover': {
+      transform: 'scale(1.02)',
+      boxShadow: '0 4px 8px rgba(0,0,0,0.1)',
+    },
+  },
+  cardHeader: {
+    display: 'flex',
+    height: '10%',
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    alignItems: 'center',
+    padding: '0.5rem',
+    '& .MuiCardHeader-action': {
+      margin: 0,
+      alignSelf: 'center',
+    },
+  },
+  followButton: (isFollowing: boolean) => ({
+    color: isFollowing ? 'black' : 'primary',
+    borderColor: isFollowing ? 'black' : 'primary',
+    mb: '0',
+  }),
+  postInfoBox: {
+    display: 'flex',
+    fontSize: '11px',
+    color: '#666',
+    gap: '5px',
+  },
+  cardMediaBox: {
+    position: 'relative',
+    height: '60%',
+    width: '100%',
+    overflow: 'hidden',
+    boxSizing: 'border-box',
+    backgroundColor: '#000',
+    zIndex: 1,
+  },
+  cardBackground: {
+    position: 'absolute',
+    width: '100%',
+    height: '100%',
+    objectFit: 'cover',
+    filter: 'blur(10px) brightness(0.5)',
+    transform: 'scale(1.1)',
+  },
+  bookImageBox: {
+    position: 'relative',
+    width: '70%',
+    height: '100%',
+    margin: '0 auto',
+    display: 'flex',
+    justifyContent: 'center',
+    alignItems: 'center',
+    boxSizing: 'border-box',
+  },
+  bookImage: {
+    width: '80%',
+    height: '80%',
+    objectFit: 'contain',
+    aspectRatio: '2/3',
+    zIndex: 3,
+    transform: 'scale(0.7)',
+  },
+  cardContent: {
+    width: '100%',
+    maxWidth: '100%',
+    overflow: 'hidden',
+    padding: '1rem 1rem 0.5rem 1rem',
+    height: '20%',
+    borderRadius: '8px 8px 0px 0px',
+    marginTop: '-15px',
+    position: 'relative',
+    backgroundColor: '#fafafa',
+    zIndex: 2,
+    boxSizing: 'border-box',
+  },
+  cardTitleBox: {
+    display: 'flex',
+    flexDirection: 'column',
+    justifyContent: 'center',
+    alignItems: 'center',
+    width: '100%',
+    overflow: 'hidden',
+  },
+  cardTitle: {
+    fontSize: '18px',
+    fontWeight: '700',
+    color: '#333',
+    width: '100%',
+    textAlign: 'center',
+    display: '-webkit-box',
+    WebkitLineClamp: 1,
+    WebkitBoxOrient: 'vertical',
+    overflow: 'hidden',
+  },
+  bookAuthor: {
+    fontSize: '12px',
+    fontWeight: '300',
+  },
+  cardDescriptionBox: {
+    width: '100%',
+    boxSizing: 'border-box',
+  },
+  cardSentence: {
+    padding: '1rem 0',
+    width: '100%',
+    height: '100%',
+    color: '#333',
+    boxSizing: 'border-box',
+    wordBreak: 'break-word',
+    whiteSpace: 'pre-wrap',
+  },
+  postingDescription: {
+    height: 'auto',
+    padding: '0.25rem 0.5rem',
+    color: '#333',
+    textOverflow: 'ellipsis',
+    overflow: 'hidden',
+    display: '-webkit-box',
+    WebkitLineClamp: 6,
+    WebkitBoxOrient: 'vertical',
+    wordBreak: 'break-all',
+    lineHeight: '20px',
+  },
+  cardFooter: {
+    padding: '0',
+    height: '10%',
+    borderTop: '1px solid #ddd',
+    gap: 0,
+    display: 'grid',
+    gridTemplateColumns: '1fr 1fr',
+    boxSizing: 'border-box',
+  },
+  cardFooterButton: (isLeft: boolean) => ({
+    padding: '0.7rem 0',
+    width: '100%',
+    height: '100%',
+    borderRadius: isLeft ? '0px 0px 0px 8px' : '0px 0px 8px 0px',
+    borderRight: isLeft ? '1px solid #ddd' : 'none',
+    bgcolor: 'transparent',
+    color: '#333',
+    transition: 'background-color 0.3s ease',
+    '&:hover': {
+      bgcolor: '#f0f0f0',
+    },
+  }),
+};
